Fix bower to not break when used in method chain, Fix jsVendor
Bower task is a stream type task. It was being wrapped in the a function to provide the destination parameter, but this is read by gulp as a normal function task.

Gulp tries to inject the callback into this and causes failure. This was masked by the old structure calling bower in a different task, with empty arguments.

jendor js relies on bower and was updated as a result too.